Cleanup pipeline file and undo feature selection
Somehow the pipeline contained duplicated fragments that I removed.
Also usingt the feature selection to unselect some Intensity features
is something I currently would not advise for the standard pipeline:
Currenlty this hard-codes the actual feature names saved. As the channel
names are part of this feature names, that means that if eg a user
modifies the pipeline to add more channels, they will per default NOT be
exported, except if the user modifies the pipeline to do so (explicitly:
push the 'select measurements' button, and manually add these channels
to the exported measurements).
Either this a) needs to be well documented or b) we do not use this
feature selection feature to avoid any confusion.